smartcity/smartcity_app/requirements.txt
asgiref==3.2.3
bleach==3.1.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
django-geolocation-fields==0.1.dev1
docutils==0.16
idna==2.8
importlib-metadata==1.4.0
jeepney==0.4.2
keyring==21.1.0
more-itertools==8.1.0
pkginfo==1.5.0.1
pycparser==2.19
Pygments==2.5.2
pytz==2019.3
readme-renderer==24.0
requests==2.22.0
requests-toolbelt==0.9.1
SecretStorage==3.1.2
six==1.14.0
sqlparse==0.3.0
tqdm==4.41.1
twine==3.1.1
urllib3==1.25.8
webencodings==0.5.1
zipp==2.0.1
